Catch UnicodeDecideError when reading file
Fixes an issue where flynt would crash with a UnicodeDecodeError if a file
contained a bad Unicode sequence. The exception didn't say which file had
a problem, making it hard to debug.

I had some trouble committing a corrupted sample file to git, so the test
generates a file instead.