remove extra newline added by sys.stdin
When running flynt on stdin, there is an extra newline added to the input, which ends up in the output.

To reproduce the issue, run this before and after this commit (after writing some dummy code to in.py):

```
cat in.py  | flynt - > out.py
diff in.py out.py
```

The only diff should be the f-strings.